+++ src/choreographer/utils/_tmpfile.py
@@ -183,7 +183,7 @@ class TmpDirectory:
             if hasattr(self, "temp_dir"):
                 del self.temp_dir
             _logger.info("shutil.rmtree worked.")
-        except Exception as e:  # noqa: BLE001
+        except Exception as e:
             _logger.debug("Error during tmp file removal.", exc_info=e)
             self._delete_manually(check_only=True)
             if not self.exists: